% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{mean_two}
\alias{mean_two}
\title{Perform a simple arithmetic mean of two values.
Parameters x and y are the values you want the mean of}
\usage{
mean_two(x, y)
}
\value{
mean value
}
\description{
Perform a simple arithmetic mean of two values.
Parameters x and y are the values you want the mean of
}
